Reactifying the log client app
