+++ src/plot/client/.gitignore
@@ -1,1 +1,1 @@
-build
+node_modules
