+++ src/plot/client/babel.config.json
@@ -1,0 +1,5 @@
+{
+    "presets": [
+        "@babel/preset-react"
+    ]
+}
